+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.6.17 (2020-04-02)
+
+- add file cache as last resort to ensure a robust api in cost of fresh data
+- add unidentifed items support
+- add armour/ evasion/ energy shield/ dps with 20% quality (#136)
+- update use same height for each mod whether selected or not (#540)
+- update evaluate search error display to include the actual reason (#537)
+- fix `Titan's Arcade Map of Temporal Chains` mismatched with `Temporal Chains` (my patrons)
+
 ## 0.6.16 (2020-03-31)
 
 - add disabled debounce time on max value (#522)